# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRMS CRM API Documentation",
        "description": "API documentation for the HRMS CRM application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://hrmscrm.test",
            "description": "Local Development Server"
        },
        {
            "url": "https://ymhrms.girafdev.com",
            "description": "Demo Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/rewards/categories": {
            "get": {
                "tags": [
                    "Rewards"
                ],
                "summary": "Get reward categories",
                "description": "Get list of active reward categories.",
                "operationId": "6a9684e7551bc5e5cef433c44d9575df",
                "responses": {
                    "200": {
                        "description": "Reward categories retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}